website: shared Button, Pill and Card components; split the asset test
The buttons, pills and cards were class names with rules in the global
stylesheet. They are now components in lib/components that own their
styles: `Button` (primary or secondary, internal or external link),
`Pill` (optionally linked and accented) and `Card` (a div, or an li for
cards in a list). The hero, the footer call to action and the feature
grid use them; the unused ghost button variant is gone. The rendered
HTML is unchanged and both routes stay pixel-identical.

The asset golden test is split into what the assets are
(test/assets_test.dart), how they are drawn (AssetRenderer, which
renders the SVG mark or any Jaspr component in the pinned Chrome) and
how they are judged (the pixelmatch-style comparison).
